Move PlatformBundles class out of 'xml' package.
