+++ config-model/src/main/java/com/yahoo/vespa/model/container/PlatformBundles.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.container.xml;
+package com.yahoo.vespa.model.container;
 
 import com.yahoo.vespa.defaults.Defaults;
